refactor: define HtmlService interface and update query logic
